+++ packages/hunspell-reader/src/de.test.ts
@@ -2,13 +2,13 @@ import { describe, expect, it } from 'vitest';
 
 import * as Aff from './affLegacy.js';
 import * as AffReader from './affReader.js';
-import { IterableHunspellReader } from './IterableHunspellReader.js';
+import { IterableHunspellReaderLegacy } from './IterableHunspellReaderLegacy.js';
 
 describe('HunspellReader DE', function () {
     it('tests transforming `Arbeit/A-`', () => {
         const affInfo = AffReader.parseAff(deAff());
         const aff = new Aff.Aff(affInfo);
-        const reader = new IterableHunspellReader({ aff, dic: ['Arbeit/A-'] });
+        const reader = new IterableHunspellReaderLegacy({ aff, dic: ['Arbeit/A-'] });
         const words = [...reader.seqWords()];
         expect(words).toEqual(['-Arbeit', '-Arbeits', 'Arbeit', 'Arbeits', 'arbeit', 'arbeits']);
     });
@@ -16,7 +16,7 @@ describe('HunspellReader DE', function () {
     it('tests transforming some entries', () => {
         const affInfo = AffReader.parseAff(deAff());
         const aff = new Aff.Aff(affInfo);
-        const reader = new IterableHunspellReader({ aff, dic: deDic() });
+        const reader = new IterableHunspellReaderLegacy({ aff, dic: deDic() });
         const words = [...reader];
         expect(words).toEqual(['-Arbeit', 'Arbeit', '-Computer', '-Computern', 'Computer', 'Computern', '-']);
     });
@@ -24,7 +24,7 @@ describe('HunspellReader DE', function () {
     it('tests transforming a small dictionary', () => {
         const affInfo = AffReader.parseAff(deAff());
         const aff = new Aff.Aff(affInfo);
-        const reader = new IterableHunspellReader({ aff, dic: deDic() });
+        const reader = new IterableHunspellReaderLegacy({ aff, dic: deDic() });
         const words = [...reader.seqAffWords()].map((w) => Aff.debug.signature(w));
         // cspell:ignore CEO BCO BCMO CMO CEMO
         expect(words).toEqual([
