+++ packages/hunspell-reader/src/nl.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 
 import * as Aff from './affLegacy.js';
 import * as AffReader from './affReader.js';
-import { IterableHunspellReader } from './IterableHunspellReader.js';
+import { IterableHunspellReaderLegacy } from './IterableHunspellReaderLegacy.js';
 
 const timeout = 10000;
 
@@ -20,7 +20,7 @@ describe('HunspellReader NL', function () {
         'tests transforming some entries',
         async () => {
             const aff = await pAff;
-            const reader = new IterableHunspellReader({ aff, dic: ['baddoek/Zb'] });
+            const reader = new IterableHunspellReaderLegacy({ aff, dic: ['baddoek/Zb'] });
             const words = [...reader];
             expect(words).toEqual(['baddoek', 'baddoeken', 'baddoeken-']);
         },
@@ -33,7 +33,7 @@ describe('HunspellReader NL', function () {
         'tests transforming some entries with debug signatures',
         async () => {
             const aff = await pAff;
-            const reader = new IterableHunspellReader({ aff, dic: ['baddoek/Zb'] });
+            const reader = new IterableHunspellReaderLegacy({ aff, dic: ['baddoek/Zb'] });
             const words = [...reader.seqAffWords()].map((w) => Aff.debug.signature(w));
             expect(words).toEqual(['baddoek|', 'baddoeken|BCM', 'baddoeken-|BCM']);
         },
